Move position_t out of mmo.h
